==> src/02_biodiv_preperation.R <==
source("C:/Users/uselig/Documents/GitHub/project_biodiv_rs/src/00_paths_libraries.R")
# the preparation of the vegrel data has been performed in the initial workflow in 2017
 vegrel15 <- readRDS(file = paste0(path_rdata_pre,"vegrel15.rds"))
